Revert "refactor(compiler): remove cyclic dependencies (#42062)"
This reverts commit 75855196e320022e79202ccaeb9be000da55b9eb.
