+++ packages/compiler/src/i18n/extractor.ts
@@ -28,6 +28,7 @@ import {NgModuleResolver} from '../ng_module_resolver';
 import {ParseError} from '../parse_util';
 import {PipeResolver} from '../pipe_resolver';
 import {DomElementSchemaRegistry} from '../schema/dom_element_schema_registry';
+import {syntaxError} from '../util';
 
 import {MessageBundle} from './message_bundle';
