+++ packages/compiler/src/render3/r3_factory.ts
@@ -5,10 +5,17 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
+
+import {StaticSymbol} from '../aot/static_symbol';
+import {CompileTypeMetadata, tokenReference} from '../compile_metadata';
 import {InjectFlags} from '../core';
 import * as o from '../output/output_ast';
 import {Identifiers as R3} from '../render3/r3_identifiers';
+import {OutputContext} from '../util';
+
 import {R3CompiledExpression, R3Reference, typeWithParameters} from './util';
+import {unsupported} from './view/util';
+
 
 
 /**
