+++ packages/compiler/src/style_compiler.ts
@@ -6,13 +6,12 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {CompileDirectiveMetadata, CompileStylesheetMetadata} from './compile_metadata';
-import {OutputContext} from './constant_pool';
+import {CompileDirectiveMetadata, CompileIdentifierMetadata, CompileStylesheetMetadata, identifierModuleUrl, identifierName} from './compile_metadata';
 import {ViewEncapsulation} from './core';
 import * as o from './output/output_ast';
-import {identifierModuleUrl, identifierName} from './parse_util';
 import {ShadowCss} from './shadow_css';
 import {UrlResolver} from './url_resolver';
+import {OutputContext} from './util';
 
 const COMPONENT_VARIABLE = '%COMP%';
 export const HOST_ATTR = `_nghost-${COMPONENT_VARIABLE}`;
